requirements.txt
annotated-types==0.6.0
anyio==3.7.1
Authlib==1.3.0
azure-common==1.1.28
azure-core==1.29.5
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.8.0
ecdsa==0.19.0
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
identity==0.3.3
idna==3.5
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
msal==1.28.0
openai==1.3.5
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyodbc==5.0.1
pypdf==3.17.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
